fix: Rerouted the pages to the landing page
