+++ backend/app/Config/Routes.php
@@ -6,19 +6,26 @@ use CodeIgniter\Router\RouteCollection;
  * @var RouteCollection $routes
  */
 $routes->get('/', 'Users::index');
+
 $routes->get('/login', 'Users::login');
+
 $routes->get('/signup', 'Users::signup');
+
 $routes->get('/moodPage', 'Users::moodPage');
+
 $routes->get('/roadMap', 'Users::roadMap');
 
+$routes->get('/admin/adminDashboard', 'Users::adminDashboard');
 
-$routes->get('admin/dashboard', 'admin::index');
-$routes->get('orders', 'admin::orders');
+$routes->get('/admin/menu', 'Users::menu');
 
+$routes->get('/admin/accounts', 'Users::accounts');
 
-
+$routes->get('/admin/orders', 'Users::orders');
 
 $routes->post('login', 'Auth::login');
+
 $routes->get('logout', 'Auth::logout');
 $routes->post('logout', 'Auth::logout');
+
 $routes->post('signup', 'Auth::signup');
