+++ backend/app/Views/components/header.php
@@ -44,11 +44,14 @@
 </style>
 
 <div class="header">
-    <a class="title" href="#">🍌Banana Odyssey</a>
+    <a class="title" href="/">🍌Banana Odyssey</a>
     <ul class="btn-header">
-        <li><a href="/">Home</a></li>
-        <li><a href="/menu">Menu</a></li>
-        <li><a href="/login">Login</a></li>
-        <li><a href="/signup">Sign Up</a></li>
+        <li><a href="<?= base_url('/admin/adminDashboard') ?>" class="hover:text-yellow-200">Dashboard</a></li>
+        <li><a href="<?= base_url('/admin/menu') ?>" class="hover:text-yellow-200">Menu</a></li>
+        <li><a href="<?= base_url('/admin/accounts') ?>" class="hover:text-yellow-200">Accounts</a></li>
+        <li><a href="<?= base_url('/admin/orders') ?>" class="hover:text-yellow-200">Orders</a></li>
+        <li><a href="<?= base_url('/login') ?>" class="hover:text-yellow-200">Log In</a></li>
+        <li><a href="<?= base_url('/signup') ?>" class="hover:text-yellow-200">Sign Up</a></li>
+
     </ul>
 </div>
